+++ src/components/Form/TextArea/TextArea.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
 import { BaseInputComponent } from "../BaseInputComponent/BaseInputComponent";
-import { InputCommonProps } from "../BaseInputComponent/constants";
 import { CommonAndHTMLProps } from "../../Element/constants";
+import { InputCommonProps } from "../BaseInputComponent/constants";
 import { InputLabelProps } from "../InputLabel/InputLabel";
 import { TextareaStyled } from "./Textarea.styled";