+++ CHANGELOG.md
@@ -1,27 +1,43 @@
+# v0.28.2
+- Fix Textarea styling bug
+- Add `nano` and `micro` text sizes.
+- Text theme tweaks
+  -  `size` object deprecated.
+  -  `size.default` is now `paras.size`.
+  -  `size.multiplier` is now `headings.multiplier`.
+
+
 # v0.27.0
 - Add props for ProgressBar to display a label and value.
+
 
 # v0.23.8
 - Remove all SCSS files.
 - Remove unused svg icon.
 - Move `theme` to `styles` folder.
 
+
 # v0.23.7
 - Use relative path everywhere.
+
 
 # v0.23.6
 - Generate typings.
 
+
 # v0.23.5
 - Upgrade all packages.
+
 
 # v0.21.2
 - Update styled components.
 - Remove unused component `EventHandler`.
 - Fix rollup dependencies.
 
+
 # v0.21.1
 - Add `onCloseButtonClick` prop in `NotificationItem`.
+
 
 # v0.21.0
 - Code cleanup.